requirements.txt
fastapi
uvicorn
python-multipart
chromadb
openai
tiktoken
python-docx
sqlite-utils
pydantic
python-dotenv
